requirements.txt
customtkinter==5.1.3
keyboard==0.13.5
matplotlib==3.7.1
mediapipe==0.9.3.0
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
openpyxl==3.1.2
Pillow==9.5.0
Pillow==9.5.0
pyserial==3.5
